[KERNEL][VFS] Current offset in file is now stored as off_t

+++ include/kernel/vfs/vfile.h
@@ -19,7 +19,7 @@ struct vfile {
     char mode;
 
     /* Current offset in the file */
-    size_t offset;
+    off_t offset;
 
     ino_t inode;
